refactor(core): group the application routes under applications directory
group the application routes under applications directory
