+++ packages/core/src/routes/applications/application-role.openapi.json
