+++ packages/core/src/routes/applications/application.openapi.json
